LanguagePicker redux methods moved to KnownLanguages
