+++ js/actions/languages.js
@@ -3,16 +3,16 @@ import type { Action } from './types';
 export const ADD_LANGUAGE = 'ADD_LANGUAGE';
 export const REMOVE_LANGUAGE = 'REMOVE_LANGUAGE';
 
-export function addLanguage(name:string):Action {
+export function addLanguage(id:string):Action {
   return {
     type: ADD_LANGUAGE,
-    payload: name,
+    payload: id,
   };
 }
 
-export function removeLanguage(name:string):Action {
+export function removeLanguage(id:string):Action {
   return {
     type: REMOVE_LANGUAGE,
-    payload: name,
+    payload: id,
   };
 }
