+++ js/components/developer/language-picker-screen/languages.js
@@ -1,0 +1,12 @@
+// Temporary array with languages
+const LANGUAGES = [{ name: 'Abkhaz', id: 1 }, { name: 'Adyghe', id: 2 },
+  { name: 'Afrikaans', id: 3 }, { name: 'Akan', id: 4 },
+  { name: 'Albanian', id: 5 }, { name: 'Amharic', id: 6 },
+  { name: 'Arabic', id: 7 }, { name: 'Aragonese', id: 8 },
+  { name: 'Aramaic', id: 9 }, { name: 'Armenian', id: 10 },
+  { name: 'Aymara', id: 11 }, { name: 'English', id: 12 },
+  { name: 'French', id: 13 }, { name: 'German', id: 14 },
+  { name: 'Mandarin', id: 15 }, { name: 'Swahili', id: 16 },
+  { name: 'Swedish', id: 17 }, { name: 'Tagalog', id: 18 }];
+
+module.exports = LANGUAGES;
